feat(core) - Add methods for renaming keys on Hash (#23)
Closes #20 

The original design approached this feature from the perspective of "I
need to replace this one key in this Hash". But I realized that the
perspective needed changed and I needed to look at this as "I want to
rename this key in this Hash".

After much benchmarking and testing, I landed on two versions based on
ordering and speed.